requirements.txt
img2pdf==0.5.1
Pillow==10.2.0
Requests==2.31.0
